requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
black==25.9.0
cffi==2.0.0
click==8.3.0
Django==5.2.7
django-axes==8.0.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyotp==2.9.0
pytokens==0.1.10
qrcode==8.2
sqlparse==0.5.3
